refactor: address daily quality findings for 2026-08-10
Extract Task GitHub create-issue template form fields and panel model
constants to shrink the new Tasks surface hubs.
